Move ProcStatus to own file.
