+++ process/plugin/plugin.go
@@ -69,23 +69,6 @@ func (p ProcDetails) validate() error {
 	return p.ProcStatus.validate()
 }
 
-// ProcStatus represents the data returned from the Status call.
-type ProcStatus struct {
-	// Status represents the human-readable string returned by the plugin for
-	// the process.
-	Status string `json:"status"`
-}
-
-// validate returns nil if this value is valid, and an error that satisfies
-// IsValid if it is not.
-func (p ProcStatus) validate() error {
-	if p.Status == "" {
-		e := errors.NewErr("Status cannot be empty")
-		return validationErr{&e}
-	}
-	return nil
-}
-
 // Plugin represents a provider for launching, destroying, and introspecting
 // workload processes via a specific technology such as Docker or systemd.
 type Plugin struct {
